lcomment | refactor | #28 | 도메인 수정으로 인한 S3 관련 기능 리팩토링
