+++ src/main/java/com/ege/wooda/global/s3/dto/S3File.java
@@ -1,4 +1,4 @@
-package com.ege.wooda.global.s3;
+package com.ege.wooda.global.s3.dto;
 
 public record S3File(String fileName, String fileUrl) {
